requirements.txt
Django==2.1.5
alabaster==0.7.12
asn1crypto==0.24.0
astroid==2.1.0
astropy==3.1.2
atomicwrites==1.3.0
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bkcharts==0.2
bleach==3.1.0
bokeh==1.0.4
boto==2.49.0
Bottleneck==1.2.1
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
colorama==0.4.1
contextlib2==0.5.5
cryptography==2.5
cycler==0.10.0
Cython==0.29.5
cytoolz==0.9.0.1
dask==1.1.3
decorator==4.3.2
defusedxml==0.5.0
distributed==1.26.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-webpack-loader==0.6.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.10
gevent==1.4.0
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.15
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.7
jsonschema==2.6.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.27.0
locket==0.2.0
lxml==4.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
numba==0.42.0
numexpr==2.6.9
odo==0.5.1
olefile==0.46
openpyxl==2.6.0
packaging==19.0
parso==0.3.4
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pkginfo==1.5.0.1
pluggy==0.9.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.5.0
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.1.0
pyparsing==2.3.1
PyPDF2==1.26.0
